requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
schedule
google-generativeai
pandas
openpyxl
xlsxwriter
streamlit
plotly
psutil
supabase
requests
openai
moviepy
Pillow
python-dotenv
watchdog
